backend/ml/requirements.txt
tensorflow-macos
tensorflow-metal
flask
pillow
flask-cors
numpy
scikit-learn 